% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{desire_lines_leeds}
\alias{desire_lines_leeds}
\title{Cycle route desire lines for Leeds}
\description{
Cycle route desire lines for Leeds
}
\examples{
# see data-raw folder for generation code
desire_lines_leeds
}
\keyword{datasets}
