% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_centroids_ew}
\alias{get_centroids_ew}
\title{Download MSOA centroids for England and Wales}
\usage{
get_centroids_ew()
}
\description{
Downloads and processes data on where people live in England and Wales.
See https://data.cdrc.ac.uk/dataset/cdrc-2011-population-weighted-centroids-gb
}
\examples{
pwc = get_centroids_ew()
plot(pwc[sample(nrow(pwc), 1000), ])
}
