% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pct.R
\name{get_pct_centroids}
\alias{get_pct_centroids}
\title{Get centroid results from the PCT}
\usage{
get_pct_centroids(region = NULL, purpose = "commute",
  geography = "msoa")
}
\arguments{
\item{region}{The PCT region that contains the data (e.g. \code{west-yorkshire})}

\item{purpose}{Trip purpose (typically \code{school} or \code{commute})}

\item{geography}{Geographic resolution of outputs (\code{msoa} or \code{lsoa})}
}
\description{
Wrapper around \code{[get_pct()]} that gets centroid data from the PCT.
}
\examples{
z = get_pct_centroids("isle-of-wight")
plot(z)
}
