% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pct.R
\name{get_pct_zones}
\alias{get_pct_zones}
\title{Get zone results from the PCT}
\usage{
get_pct_zones(region = NULL,
  base_url = "https://github.com/npct/pct-outputs-regional-R/raw/master",
  purpose = "commute", geography = "msoa", layer = "z",
  extension = ".Rds")
}
\arguments{
\item{region}{The PCT region that contains the data (e.g. \code{west-yorkshire})}

\item{base_url}{Where the data is stored.}

\item{purpose}{Trip purpose (typically \code{school} or \code{commute})}

\item{geography}{Geographic resolution of outputs (\code{msoa} or \code{lsoa})}

\item{layer}{The PCT layer of interest, \code{z}, \code{c}, \code{l}, \code{rf}, \code{rq} or \code{rnet}
for zones, centroids, desire lines, routes (fast or quiet) and route networks, respectively}
}
\description{
Wrapper around \code{[get_pct()]} that gets zone data from the PCT.
}
\examples{
z = get_pct_zones("isle-of-wight")
plot(z)
}
