% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leeds_uber_sample}
\alias{leeds_uber_sample}
\title{Top 15 min mean journy times within Leeds from Uber}
\description{
Top 15 min mean journy times within Leeds from Uber
}
\examples{
# see data-raw folder for generation code
leeds_uber_sample
}
\keyword{datasets}
