% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{od_leeds}
\alias{od_leeds}
\title{Example OD data for Leeds}
\description{
\code{od_leeds} contains the 100 most travelled work desire lines in Leeds,
according to the 2011 Census.
}
\examples{
# see data-raw folder for generation code
od_leeds
}
\keyword{datasets}
