% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desire_lines.R
\name{pct_area_desire_lines}
\alias{pct_area_desire_lines}
\title{Desire lines}
\usage{
pct_area_desire_lines(area = "sheffield", n = 100)
}
\arguments{
\item{area}{for which desire lines to be generated.}

\item{n}{top n number of destinations with most trips in the 2011 census
within the \code{area}.}
}
\description{
Desire lines
}
