% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pct_lines_min}
\alias{pct_lines_min}
\title{A minimum example of desire line data from the PCT}
\description{
This data was obtained using code shown in the introductory
\href{https://itsleeds.github.io/pct/articles/pct.html}{pct package vignette}.
}
\examples{
names(pct_lines_min)
plot(pct_lines_min)
}
\keyword{datasets}
