% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pct_regions_lookup}
\alias{pct_regions_lookup}
\title{Lookup table matching PCT regions to local authorities}
\description{
For matching pct_regions object with local authority names in England and Wales.
}
\examples{

names(pct_regions_lookup)
head(pct_regions_lookup)
}
\keyword{datasets}
