% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{routes_fast_leeds}
\alias{routes_fast_leeds}
\title{Fastest cycle routes for the desire_lines_leeds}
\description{
Fastest cycle routes for the desire_lines_leeds
}
\examples{
# see data-raw folder for generation code
routes_fast_leeds
}
\keyword{datasets}
