% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{santiago_lines}
\alias{santiago_lines}
\title{Desire lines in central Santiago}
\description{
See https://github.com/pedalea/pctSantiago folder for generation code
}
\examples{
# u = "https://github.com/pedalea/pctSantiago/releases/download/0.0.1/od_agg_zone_sub.Rds"
# download.file(u, destfile = "od_agg_zone_sub.Rds")
# desire_lines = readRDS("od_agg_zone_sub.Rds")
santiago_zones
}
\keyword{datasets}
