% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{santiago_od}
\alias{santiago_od}
\title{OD data in central Santiago}
\description{
See https://github.com/pedalea/pctSantiago folder for generation code
}
\examples{
# u = "https://github.com/pedalea/pctSantiago/releases/download/0.0.1/santiago_od.Rds"
# download.file(u, destfile = "santiago_od.Rds")
# santiago_od = readRDS("santiago_od.Rds")
santiago_od
}
\keyword{datasets}
