% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{santiago_routes_cs}
\alias{santiago_routes_cs}
\title{200 cycle routes in central Santiago, Chile}
\description{
This data was obtained using code shown in the
International application of the PCT methods
\href{https://itsleeds.github.io/pct/articles/pct-international.html}{vignette}.
}
\examples{
names(santiago_routes_cs)
head(santiago_routes_cs)
plot(santiago_routes_cs)
}
\keyword{datasets}
