% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{santiago_zones}
\alias{santiago_zones}
\title{Zones in central Santiago}
\description{
See https://github.com/pedalea/pctSantiago folder for generation code
}
\examples{
# u = "https://github.com/pedalea/pctSantiago/releases/download/0.0.1/z_centre.Rds"
# download.file(u, destfile = "z_centre.Rds", mode = "wb")
# santiago_zones = readRDS("z_centre.Rds")
santiago_zones
}
\keyword{datasets}
