% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ukboundaries}
\alias{ukboundaries}
\title{UK MSOA boundary codes and their names in plain text}
\description{
UK MSOA boundary codes and their names in plain text
}
\examples{
# see data-raw folder for generation code
ukboundaries
}
\keyword{datasets}
