% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uptake.R
\name{uptake_pct_godutch}
\alias{uptake_pct_godutch}
\title{Go Dutch}
\usage{
uptake_pct_godutch(distance, gradient, alpha = -3.959, d1 = -0.5963,
  d2 = 1.832, d3 = 0.007956, h1 = -0.2872, i1 = 0.01784,
  i2 = -0.0977)
}
\description{
Go Dutch
}
