% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wight_lines_30}
\alias{wight_lines_30}
\alias{wight_lines_pct}
\title{Desire lines from the PCT for the Isle of Wight}
\description{
This data was obtained using code shown in the introductory
\href{https://itsleeds.github.io/pct/articles/pct.html}{pct package vignette}.
}
\examples{
names(wight_lines_30)
plot(wight_lines_30)
}
\keyword{datasets}
