% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wight_od}
\alias{wight_od}
\title{Official origin-destination data for the Isle of Wight}
\description{
This data was obtained using code shown in the introductory
\href{https://itsleeds.github.io/pct/articles/pct.html}{pct package vignette}.
}
\examples{
names(wight_od)
head(wight_od)
}
\keyword{datasets}
