% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wight_routes_30}
\alias{wight_routes_30}
\alias{wight_rnet}
\alias{wight_routes_30_cs}
\title{Cycle route data for the Isle of Wight}
\description{
This data was obtained using code shown in the introductory
\href{https://itsleeds.github.io/pct/articles/pct.html}{pct package vignette}.
}
\examples{
library(sf)
names(wight_routes_30)
head(wight_routes_30)
plot(wight_routes_30)
}
\keyword{datasets}
