% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wight_zones}
\alias{wight_zones}
\alias{wight_centroids}
\title{Zones and centroid data from the PCT for the Isle of Wight}
\description{
This data was obtained using code shown in the introductory
\href{https://itsleeds.github.io/pct/articles/pct.html}{pct package vignette}.
}
\examples{
library(sf)
names(wight_lines_30)
plot(wight_lines_30)
}
\keyword{datasets}
