% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msoa_centroids.R
\name{get_centroids_ew}
\alias{get_centroids_ew}
\title{Download MSOA centroids for England and Wales}
\usage{
get_centroids_ew()
}
\description{
Downloads and processes data on where people live in England and Wales.
See \href{http://geoportal.statistics.gov.uk/datasets/b0a6d8a3dc5d4718b3fd62c548d60f81_0}{geoportal.statistics.gov.uk}.
}
\examples{
pwc = get_centroids_ew()
plot(pwc[sample(nrow(pwc), 1000), ])
}
