% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pct.R
\name{get_pct_centroids}
\alias{get_pct_centroids}
\title{Get centroid results from the PCT}
\usage{
get_pct_centroids(region = NULL, purpose = "commute",
  geography = "msoa", extension = ".Rds")
}
\arguments{
\item{region}{The PCT region or local authority to download data from (e.g. \code{west-yorkshire} or \code{Leeds}).
See \code{View(pct_regions_lookup)} for a full list of possible region names.}

\item{purpose}{Trip purpose (typically \code{school} or \code{commute})}

\item{geography}{Geographic resolution of outputs (\code{msoa} or \code{lsoa})}

\item{extension}{The type of file to download (typically \code{.Rds})}
}
\description{
Wrapper around \code{[get_pct()]} that gets centroid data from the PCT.
}
\examples{
z = get_pct_centroids("isle-of-wight")
plot(z)
}
