% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pct.R
\name{get_pct_lines}
\alias{get_pct_lines}
\title{Get desire lines results from the PCT}
\usage{
get_pct_lines(region = NULL, purpose = "commute", geography = "msoa",
  extension = ".Rds")
}
\arguments{
\item{region}{The PCT region that contains the data (e.g. \code{west-yorkshire})}

\item{purpose}{Trip purpose (typically \code{school} or \code{commute})}

\item{geography}{Geographic resolution of outputs (\code{msoa} or \code{lsoa})}

\item{extension}{The type of file to download (typically \code{.Rds})}
}
\description{
Wrapper around \code{[get_pct()]} that gets l (lines) data from the PCT.
}
\examples{
z =  get_pct_lines("isle-of-wight")
plot(z)
}
