% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uptake.R
\name{uptake_pct_godutch}
\alias{uptake_pct_godutch}
\title{Calculate cycling uptake for UK 'Go Dutch' scenario}
\usage{
uptake_pct_godutch(
  distance,
  gradient,
  alpha = -3.959,
  d1 = -0.5963,
  d2 = 1.832,
  d3 = 0.007956,
  h1 = -0.2872,
  i1 = 0.01784,
  i2 = -0.0977
)
}
\arguments{
\item{distance}{Vector distance numeric values of routes.}

\item{gradient}{Vector gradient numeric values of routes.}

\item{alpha}{The intercept}

\item{d1}{Distance term 1}

\item{d2}{Distance term 2}

\item{d3}{Distance term 3}

\item{h1}{Hilliness term 1}

\item{i1}{document!}

\item{i2}{document!}
}
\description{
See \code{\link[=uptake_pct_govtarget]{uptake_pct_govtarget()}}.
}
\examples{
l = routes_fast_leeds
pcycle_scenario = uptake_pct_godutch(l$length, l$av_incline)
plot(l$length, pcycle_scenario)
}
