% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uptake.R
\name{uptake_pct_govtarget}
\alias{uptake_pct_govtarget}
\title{Uptake scenario for OD?}
\usage{
uptake_pct_govtarget(distance, gradient, alpha = -3.959, d1 = -0.5963,
  d2 = 1.832, d3 = 0.007956, h1 = -0.2872, i1 = 0.01784,
  i2 = -0.0977)
}
\arguments{
\item{distance}{Vector distance numeric values of routes.}

\item{gradient}{Vector gradient numeric values of routes.}

\item{d1}{document!}

\item{i2}{}
}
\description{
Uptake model that takes distance and hilliness and returns
a percentage of people likely to cycle along a desire line.
Source: appendix of pct paper:
https://www.jtlu.org/index.php/jtlu/article/downloadSuppFile/862/360
}
\examples{
l = routes_fast_leeds
pcycle_scenario = uptake_pct_govtarget(l$length, l$cum_hill)
plot(l$length, pcycle_scenario)
}
